Рефакторинг.
