+++ app/src/main/java/com/example/phasig/presentation/OptionalTimePickerState.kt
@@ -1,0 +1,28 @@
+package com.example.phasig.presentation
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+
+class OptionalTimePickerState(
+    initiallySelectedOptionH: Int = 0,
+    initiallySelectedOptionM: Int = 0,
+    initialEnabled: Boolean = false
+)
+{
+    val timePickerState: TimePickerState
+    var tpkrEnabled by mutableStateOf(initialEnabled)
+
+    init {
+        timePickerState = TimePickerState(initiallySelectedOptionH, initiallySelectedOptionM)
+    }
+}
+
+@Composable
+fun rememberOptionalTimePickerState(
+    initiallySelectedOptionH: Int,
+    initiallySelectedOptionM: Int,
+    initialEnabled: Boolean
+) = remember { OptionalTimePickerState(initiallySelectedOptionH, initiallySelectedOptionM, initialEnabled) }
